+++ app/src/SightingsList.js
@@ -1,11 +1,6 @@
 import 'bootstrap/dist/css/bootstrap.min.css';
-
-import { Dropdown, MenuItem, DropdownButton } from "react-bootstrap";
 import React, { useState, useEffect, useCallback } from "react";
-require('react-dom');
-window.React2 = require('react');
-console.log(window.React1 === window.React2);
-console.log(window.React1, window.React2);
+import FilterSort from './FilterSort';
 // function Sighting({sighting}) {
 //   return (
 //     <tr className='row'>
@@ -28,7 +23,7 @@ function SightingsList() {
   const [timestamp, setTimestamp] = useState('');
   const [sightings, setSightings] = useState([]);
   const [permanentSightings, setPermanentSightings] = useState([]);
-  const [filterNickname, setFilterNickname] = useState('');
+  
   const getSightings = () => {
     fetch("http://localhost:3000/sightings")
       .then(res => {
@@ -59,9 +54,7 @@ function SightingsList() {
         return res.json()
       })
       .then((res) => {
-        // const newSightings = [...sightings, res];
         setSightings(res);
-        // setSightings(newSightings);
       })
   };
   useEffect(() => {
@@ -80,63 +73,10 @@ function SightingsList() {
     setEmail('');
     setTimestamp('');
   };
-  const filter = (filterNicknameParam) => {
-    // e.preventDefault();
-    setSightings(permanentSightings.filter(function(sighting) {
-      console.log("sigting - nickName",sighting.nickName);
-      console.log("filter-nickname-prarm", filterNicknameParam);
-      return (sighting.nickname.includes(filterNicknameParam));
-    }))
-    console.log("sightings", sightings);
-    console.log("filter Nickname Param", filterNicknameParam);
-    console.log("filter Nickname", filterNickname);
-  };
-  const onChange = (e) => {
-    setFilterNickname(e.target.value);
-    filter(e.target.value);
-    console.log("console", e.target.value);
-  };
   return (
     <div>
       <h3>Sighting Log</h3>
-      <div className='container'>
-        <div className='row'>
-          <div className='col-sm-3'></div>
-          <div className='col-sm-3'></div>
-          <div className='col-sm-3'>
-            <Dropdown>
-              <Dropdown.Toggle variant="success" id="dropdown-basic" className='button'>
-                Filter By
-              </Dropdown.Toggle>
-
-              <Dropdown.Menu>
-                <Dropdown.Item href="#/action-1">Date</Dropdown.Item>
-                <input placeholder='Nickname' value={filterNickname} onChange= {onChange}></input>
-                <Dropdown.Item href="#/action-3">Location</Dropdown.Item>
-                <Dropdown.Item href="#/action-1">Health status</Dropdown.Item>
-                <Dropdown.Item href="#/action-2">Email</Dropdown.Item>
-                <Dropdown.Item href="#/action-3">Record date</Dropdown.Item>
-              </Dropdown.Menu>
-            </Dropdown>
-          </div>
-          <div className='col-sm-3'>
-            <Dropdown>
-              <Dropdown.Toggle variant="success" id="dropdown-basic" className='button'>
-                Sort By
-              </Dropdown.Toggle>
-
-              <Dropdown.Menu>
-              <Dropdown.Item href="#/action-1">Date</Dropdown.Item>
-                <Dropdown.Item href="#/action-2">Nickname</Dropdown.Item>
-                <Dropdown.Item href="#/action-3">Location</Dropdown.Item>
-                <Dropdown.Item href="#/action-1">Health status</Dropdown.Item>
-                <Dropdown.Item href="#/action-2">Email</Dropdown.Item>
-                <Dropdown.Item href="#/action-3">Record date</Dropdown.Item>
-              </Dropdown.Menu>
-            </Dropdown>
-          </div>
-        </div>
-      </div>
+      <FilterSort permanentSightings={permanentSightings} setSightings={setSightings}/>
       <div className='table-responsive'>
         <table className='table table-hover'>
             <thead>
